third_party/blink/web_tests/external/wpt/fetch/private-network-access/service-worker.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failure. assert_equals: error expected (string) "TypeError" but got (undefined) undefined
FAIL treat-as-public to private: failure. assert_equals: error expected (string) "TypeError" but got (undefined) undefined
PASS public to public: success.
Harness: the test ran to completion.
